+++ singer-connectors/target-snowflake/target_snowflake/upload_clients/s3_upload_client.py
@@ -102,6 +102,16 @@ class S3UploadClient(BaseUploadClient):
         source_bucket, source_key = copy_source.split("/", 1)
         metadata = self.s3_client.head_object(Bucket=source_bucket, Key=source_key).get('Metadata', {})
         metadata.update(target_metadata)
+        copy_args = {}
+        s3_acl = self.connection_config.get('s3_acl')
+        if s3_acl:
+            copy_args['ACL'] = s3_acl
         # https://boto3.amazonaws.com/v1/documentation/api/latest/reference/services/s3.html#S3.Client.copy_object
-        self.s3_client.copy_object(CopySource=copy_source, Bucket=target_bucket, Key=target_key,
-                                   Metadata=metadata, MetadataDirective="REPLACE")
+        self.s3_client.copy_object(
+            CopySource=copy_source,
+            Bucket=target_bucket,
+            Key=target_key,
+            Metadata=metadata,
+            MetadataDirective="REPLACE",
+            **copy_args,
+        )